Refactor: extracted properties stuff from main controller
