+++ src/main/java/cmpt276/pg6/realtorest/controllers/UserController.java
@@ -112,7 +112,6 @@ public class UserController {
         List<User> userList = userRepo.findByEmailAndPassword(email, password);
         if (userList.isEmpty()) {
             // If no user that matches the email and password is found, return to the login page
-            // TODO Add a message to the login page that says "Invalid email or password"
             model.addAttribute("errorMessage", "Invalid Credentials Entered.");
             return "users/login";
         } else {
