+++ src/test/java/cmpt276/pg6/realtorest/controllers/MainControllerTest.java
@@ -143,7 +143,7 @@ public class MainControllerTest {
         controller.setUserRepo(userRepo);
 
         // Act
-        String result = controller.getFavourites(request, session, model);
+        String result = controller.showFavourites(request, session, model);
 
         // Assert
         assertEquals("favourites", result);
